Fixes for stacking + general cleanup
When releasing an object it's position is no longer updated, as it was unnecessary and created more problems. Instead gravity does the job.

Also resets the velocity of the object when releasing the object, as to prevent it from crashing into the ground / other objects.

Clean up consists of a few regions, comments, early returns and replacing expensive null checks with implicit bools :^)